refactor update status and create event
