+++ src/hoprd_service.rs
@@ -57,7 +57,7 @@ pub async fn delete_service(client: Client, name: &str, namespace: &str) -> Resu
         let uid = service.metadata.uid.unwrap();        
         api.delete(name, &DeleteParams::default()).await?;
         await_condition(api, &name.to_owned(), conditions::is_deleted(&uid)).await.unwrap();
-        Ok(println!("[INFO] Service successfully deleted"))
+        Ok(println!("[INFO] Service {name} successfully deleted"))
     } else {
         Ok(println!("[INFO] Service {name} in namespace {namespace} about to delete not found"))
     }
